+++ tests/export-job.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const { close, init } = require('../index');
+const { close, init } = require('./index');
 
 /**
  * test creates a new dummy ExportJob instance and tests the job.process()
@@ -8,7 +8,7 @@ const { close, init } = require('../index');
 
 test.before(async t => {
     await init();
-    const { ExportJob } = require('../../models');
+    const { ExportJob } = require('../models');
     // create new test job
     t.context = await ExportJob.create({
         key: 'test-task',
@@ -41,7 +41,7 @@ test('log progress', async t => {
     t.is(typeof t.context.log.progress, 'object');
     t.is(t.context.log.progress.length, 1);
     t.is(t.context.log.progress[0].message, 'foo');
-    t.truthy(Date.prototype.isPrototypeOf(t.context.log.progress[0].timestamp));
+    t.truthy(t.context.log.progress[0].timestamp instanceof Date);
     await t.context.logProgress({ message: 'another message' });
     t.is(t.context.log.progress.length, 2);
 });
